### 2019-10-02 #1780 @kyleniemeyer
Hi @drvinceknight, how do you feel about taking on another submission as editor?
