### 2019-10-02 #1780 @kyleniemeyer
@arfon can papers use SVG images, or does this need to be converted to (e.g.) PDF?
