### 2019-10-02 #1780 @arfon
> @arfon can papers use SVG images, or does this need to be converted to (e.g.) PDF?

Unfortunately we don't currently support SVG. We can support PNG, PDF, JPEG, etc.
