### 2019-10-03 #1780 @kyleniemeyer
@tobydriscoll could you replace the SVG image included in your article with a different format?
