### 2019-10-03 #1780 @drvinceknight
> Hi @drvinceknight, how do you feel about taking on another submission as editor?

Yup no problem.
