### 2019-10-03 149b8fa
Converted fig to pdf for paper.md
