### 2019-10-03 #1780 @tobydriscoll
I've pushed a change of the figure to PDF.

Possible reviewers are dlfivefifty, ajt60gaibb, daanhb. nakatsukasayuji.
