### 2019-10-03 5f6eb62
Fixed typos in paper references
