+++ docs/paper.bib
@@ -2,7 +2,7 @@
 
 @Article{BG87,
   author  = {P. Bj{\o}rstad and E. Grosse},
-  journal = sissc,
+  journal = {SIAM J. Sci. Stat. Comput.},
   title   = {Conformal mapping of circular-arc polygons},
   year    = {1987},
   pages   = {19--32},
@@ -24,7 +24,7 @@
 
 @Article{DE93,
   author  = {T. K. De{L}illo and A. R. Elcrat},
-  journal = j-comp-phys,
+  journal = {J. Comput. Phys.},
   title   = {Numerical Conformal Mapping Methods for Exterior Regions with Corners},
   year    = {1993},
   pages   = {199--208},
@@ -73,17 +73,9 @@
   groups  = {temp},
 }
 
-@Article{GS,
-  author = {D. Gilbarg and J. Serrin},
-  title  = {Free boundaries and jets in the theory of cavitation},
-  pages  = {1--12},
-  annote = {The wake of an object wants to turn back and form a jet crossing "over" the object. This is interpreted as existing on a different Riemann sheet, requiring the appropriately generalized SC map for free-boundary problems. Analysis is done for an inclined plate.},
-  groups = {temp},
-}
-
 @Article{Gre90,
   author  = {L. Greengard},
-  journal = j-siam-sci-stat-comp,
+  journal = {SIAM J. Sci. Stat. Comput.},
   title   = {Potential flow in channels},
   year    = {1990},
   pages   = {603--620},
@@ -115,7 +107,7 @@
 
 @Article{KT86,
   author  = {N. Kerzman and M. R. Trummer},
-  journal = jcam,
+  journal = {J. Comput. Appl. Math.},
   title   = {Numerical conformal mapping via the {S}zeg{\H{o}} kernel},
   year    = {1986},
   pages   = {125--142},
@@ -135,7 +127,7 @@
 
 @Article{OR89,
   author  = {S. T. O'Donnell and V. Rokhlin},
-  journal = sissc,
+  journal = {SIAM J. Sci. Stat. Comput.},
   title   = {A Fast Algorithm for the Numerical Evaluation of Conformal Mappings},
   year    = {1989},
   pages   = {475--487},
@@ -146,7 +138,7 @@
 
 @Article{Tre80,
   author  = {L. N. Trefethen},
-  journal = j-siam-sci-stat-comp,
+  journal = {SIAM J. Sci. Stat. Comput.},
   title   = {Numerical Computation of the {S}chwarz--{C}hristoffel Transformation},
   year    = {1980},
   pages   = {82--102},
@@ -157,7 +149,7 @@
 
 @Article{TW86,
   author  = {L. N. Trefethen and R. J. Williams},
-  journal = jcam,
+  journal = {J. Comput. Appl. Math.},
   title   = {Conformal mapping solution of {L}aplace's equation on a polygon with oblique derivative boundary conditions},
   year    = {1986},
   pages   = {227--249},
@@ -189,9 +181,9 @@
 
 @Article{Marshall2007,
   author = {Donald E. Marshall and Steffen Rohde},
-  title  = {Convergence of a Variant of the Zipper Algorithm for Conformal Mapping},
+  title  = {Convergence of a Variant of the {Z}ipper Algorithm for Conformal Mapping},
   year   = {2007},
-  issn   = {0036-1429},
+  journal= {SIAM J. Numer. Anal.},
   pages  = {2577-2609},
   volume = {45},
   doi    = {10.1137/060659119},
@@ -199,7 +191,7 @@
 
 @Article{Crowdy2016,
   author = {D. G. Crowdy and E. H. Kropf and C. C. Green and M. M. S. Nasser},
[…]
