+++ docs/paper.md
@@ -17,7 +17,7 @@ bibliography: paper.bib
 
 # Summary
 
-Complex variables offer a unique and powerful way to represent planar regions and maps between them. Since the work of Riemann and others in the 19th century, complex variables have been used to derive quantities and properties of importance to a wide range of applications in science and engineering, in addition to their inherent mathematical importance. The advent of computing accelerated and extended this trend to applications in fluids, electromagnetics, queueing theory, computer graphics, and many other fields; e.g., [@DE92;@Dri94;@Gai79;@GS;@Gre90;@MBH92;@TW86;@Ver82;@Gu2004].
+Complex variables offer a unique and powerful way to represent planar regions and maps between them. Since the work of Riemann and others in the 19th century, complex variables have been used to derive quantities and properties of importance to a wide range of applications in science and engineering, in addition to their inherent mathematical importance. The advent of computing accelerated and extended this trend to applications in fluids, electromagnetics, queueing theory, computer graphics, and many other fields; e.g., [@DE92;@Dri94;@Gai79;@Gre90;@MBH92;@TW86;@Ver82;@Gu2004].
 
 Software packages for working with complex variables, particularly conformal maps, have been in the public domain for decades [@BG87;@Crowdy2016;@Dri96;@HP83;@Hu95;@Marshall2007;@Tre80]. There are additional papers based on research codes that apparently have not made it into the public domain, e.g. [@DE93;@KT86;@OR89]. These packages tend to take the traditional form of library routines that require some degree of facility with custom data structures used to represent geometry and functions. Moreover, computational methods are mostly specialized to particular types or classes of regions, leading to parallel and duplicated functionality within a package, and incompatible representations between them.
 
