### 2019-10-03 #1780 @tobydriscoll
Checked the proof and fixed typos in the references.
