### 2019-10-07 #1780 @drvinceknight
@dlfivefifty, @ajt60gaibb, @daanhb would you be able to review this paper for JOSS? It describes a Julia package for working with complex regions.

If you have not reviewed for JOSS before, there are instructions here https://joss.readthedocs.io/en/latest/index.html and I'd be happy to answer any questions you might have.

Once we have 2 reviewers I will open a second issue where the review itself will take place.

One particular thing I'd appreciate your opinions on is the statement of need which I don't feel is made as clear as it could be in the paper
