### 2019-10-07 #1780 @drvinceknight
A few weeks would be great but if that's not possible let me know what you could possibly manage. (In my experience reviewing for JOSS takes less time and is more enjoyable than traditional publications - just FYI in case you haven't reviewed for JOSS before.)
