### 2019-10-11 #1780 @drvinceknight
Hi @vchuravy I know this doesn't fit immediately within your area of expertise but would you be willing to review this paper for JOSS?

@ChrisRackauckas, @Datseris would you be able to review this paper for JOSS?
