### 2019-10-11 #1780 @daanhb
Hi @drvinceknight, if you're looking for more reviewers, I can still chime in - though I'll probably have a similar point of view as @dlfivefifty (being co-authors of the DomainSets.jl package)
