### 2019-10-14 #1780 @drvinceknight
> Hello, thank you for the invitation, but I have to kindly decline reviewing this, as I believe it does not fall into my area of expertise.

> I am way oversubscribed until Spring. If you're still looking for reviews then, feel free to ping me.

I understand @ChrisRackauckas and @Datseris  , thank you for responding.

> @dpsanders might be a better fit than me.

Thank you @vchuravy, @dpsanders would you be able to review this paper please?

> Hi @drvinceknight, if you're looking for more reviewers, I can still chime in - though I'll probably have a similar point of view as @dlfivefifty (being co-authors of the DomainSets.jl package)

Thank you @daanhb, I'm going to wait a day or so to see if @dpsanders responds positively to gain a bit more variety of review but if not and you're still able I would be very grateful!
