### 2019-10-15 #1811 @whedon
[REVIEW]: ComplexRegions.jl: A Julia package for regions in the complex plane
**Submitting author:** @tobydriscoll (<a href="http://orcid.org/0000-0002-1490-2545">Tobin Driscoll</a>)
**Repository:** <a href="https://github.com/complexvariables/ComplexRegions.jl" target ="_blank">https://github.com/complexvariables/ComplexRegions.jl</a>
**Version:** v0.1.1
