### 2019-10-15 #1780 @drvinceknight
@dpsanders and @dlfivefifty the review will take place on #1811, please let me know if you have any questions about the process and thank you again for agreeing to take this on. It's a vital part of the process.

@daanhb thank you very much for offering to review, it's truly appreciated and hopefully you won't mind me calling on you in the future.
