### 2019-10-15 complexvariables/ComplexRegions.jl#1 @dlfivefifty
Make connection with DomainSets.jl
This package is very similar in spirit to IntervalSets.jl and DomainSets.jl. It would be good to try to combine the two as much as possible, and reference these in the paper.

One key thing: DomainSets.jl views domains/sets as an "interface" that support the `in` command. It would be good to override `in` as appropriate, e.g., the following should work:
```julia
julia> c
Circle{Complex{Float64}} in the complex plane:
   centered at (1.0 + 0.0im) with radius 1.0, negatively oriented

julia> 2.0 in c
ERROR: MethodError: no method matching iterate(::Circle{Complex{Float64}})
Closest candidates are:
  iterate(::Core.SimpleVector) at essentials.jl:604
  iterate(::Core.SimpleVector, ::Any) at essentials.jl:604
  iterate(::ExponentialBackOff) at error.jl:214
  ...
Stacktrace:
 [1] in(::Float64, ::Circle{Complex{Float64}}) at ./operators.jl:1055
 [2] top-level scope at REPL[11]:1
```
