### 2019-10-15 #1811 @dlfivefifty
I think my review is essentially done. The main thing is referencing related work of IntervalSets.jl and DomainSets.jl, see

https://github.com/complexvariables/ComplexRegions.jl/issues/1
