### 2019-10-22 complexvariables/ComplexRegions.jl#2 @dpsanders
Documentation fixes
I suggest adding an explicit `using ComplexRegions` to the first example in the docs.

Also remove / fix this:
```
/home/travis/.julia/packages/GR/ZI5OE/src/../deps/gr/bin/gksqt: error while loading shared libraries: libQt5Widgets.so.5: cannot open shared object file: No such file or directory
connect: Connection refused
GKS: can't connect to GKS socket application
Did you start 'gksqt'?

GKS: Open failed in routine OPEN_WS
GKS: GKS not in proper state. GKS must be either in the state WSOP or WSAC in routine AC
```

Part of JOSS review @ https://github.com/openjournals/joss-reviews/issues/1811
