### 2019-10-22 complexvariables/ComplexRegions.jl#2 @dpsanders
And `using Plots`.

Basically the examples should be copyable-and-pastable.
