### 2019-10-22 complexvariables/ComplexRegions.jl#2 @dpsanders
Remove `;` from `plot!(c)` so that the plot actually appears.
