### 2019-10-22 complexvariables/ComplexRegions.jl#3 @dpsanders
Meaning of `/` -- use broadcast instead?
From the point of view of Julia, it seems strange to define e.g. `1 /  ℓ`, where `ℓ` is a line.
Since  `ℓ` is a set, it would maybe be more natural to think about the broadcasted version, `1 ./  ℓ`.

Part of JOSS review @ https://github.com/openjournals/joss-reviews/issues/1811
