### 2019-10-22 complexvariables/ComplexRegions.jl#4 @dpsanders
Paper modifications
I suggest adding explicit reproducible code to make the plot in the paper.

It would also be useful to add some comments about possible use cases for this package. What downstream need does it address?

(It's a very nice package for exploring these mathematical objects, though!)

Part of JOSS review @ https://github.com/openjournals/joss-reviews/issues/1811
