### 2019-10-22 #1811 @dpsanders
My main review is now complete.

There is basically nothing written down about how to contribute to the package, but I think it's understood this is done initially via issues, which seems fine to me.

I opened a couple of issues in the repo with suggestions for improvements.
