### 2019-10-29 complexvariables/ComplexRegions.jl#1 @tobydriscoll
I will add a reference, as I was previously unaware of these two packages. I was caught up in trying to figure out the extent of domain support in `ApproxFun`, which seems related but more limited...?

The code uses `in` for membership in regions, but it's not supported for curves and paths. I'm not sure if you mean set membership or geometric membership. The former is ill-posed in floating point, and the latter changes the meaning of set membership attached to `in`.

```julia
 julia> 0 in unitdisk
true

julia> isinside(0,Circle(0,1))
true
```

Arguably I'm being inconsistent since I use "soft" tolerances for a lot of things, including region membership, but somehow the question of detecting whether a point is on a curve feels egregiously bad to me.
