### 2019-10-29 complexvariables/ComplexRegions.jl#1 @tobydriscoll
Addressed the reference issue in 94a7c29.
