### 2019-10-29 complexvariables/ComplexRegions.jl#3 @tobydriscoll
I think of broadcasting as applying to iterable collections.

An alternative analogy would be a vector. Obviously `2v` and `2.*v` have identical effects on a `Vector`, but the former holds up as scaling a vector in infinite-dimensional space, while the notion of broadcasting becomes shakier. Conversely, `v.+2` is required over `v+2` because the notion of adding vector and scalar is not a strong one conceptually.
