### 2019-10-29 complexvariables/ComplexRegions.jl#3 @dpsanders
I see what you mean, but I am thinking about it as "apply the operation to all elements of the *set*".

For example, does it make sense to define a function like `f(z) = z^2 + 1 / z`  and then apply it to a `Line`?
