### 2019-10-31 complexvariables/ComplexRegions.jl#3 @tobydriscoll
It's not unusual in complex analysis to speak of transformations mapping curves to curves, or regions to regions. Perhaps it's because of the outsize role played by conformal maps.

When I use `broadcast`, I expect it to work for any scalar-valued function requiring a scalar, including functions I define myself. That gets really complicated here. For instance, the `abs` function maps scalar to scalar, but in aggregate it maps a complex contour to some multiply-covered interval on the real line. At a minimum I'd have to dispatch `broadcast` on curve and region types, examine the function being broadcast, and error out if it wasn't on the approved list. I assume it could be done, but it sounds very clunky. And I don't know how all this is supposed to interact with the idea of dot fusion.

I don't see what is to be gained here. If you can point me to an example of a broadcast operator being used on uncountable collections, then I would better grasp how it would all work.
