### 2019-10-31 complexvariables/ComplexRegions.jl#3 @dpsanders
Yes I see your point. OK, it was probably ill-considered on my part.
Thanks for the discussion.
