### 2019-11-01 8553206
Responding to comments in issue #4
