+++ docs/paper.bib
@@ -73,6 +73,22 @@
   groups  = {temp},
 }
 
+@article{GT19,
+  title = {Solving {L}aplace Problems with Corner Singularities via Rational Functions},
+  volume = {57},
+  issn = {0036-1429, 1095-7170},
+  abstract = {A new method is introduced for solving Laplace problems on two-dimensional regions with corners by approximation of boundary data by the real part of a rational function with fixed poles exponentially clustered near each corner. Greatly extending a result of D. J. Newman in 1964 in approximation theory, we first prove that such approximations can achieve root-exponential convergence for a wide range of problems, all the way up to the corner singularities. We then develop a numerical method to compute approximations via linear least-squares fitting on the boundary. Typical problems are solved in {$<$} 1s on a desktop to 8-digit accuracy, with the accuracy guaranteed in the interior by the maximum principle. The computed solution is represented globally by a single formula, which can be evaluated in a few microseconds at each point.},
+  language = {en},
+  number = {5},
+  journal = {SIAM Journal on Numerical Analysis},
+  doi = {10/gf9fzf},
+  author = {Gopal, Abinand and Trefethen, Lloyd N.},
+  month = jan,
+  year = {2019},
+  pages = {2074-2094},
+  file = {/home/driscoll/Zotero/storage/28GXJH84/Gopal and Trefethen - 2019 - Solving Laplace Problems with Corner Singularities.pdf}
+}
+
 @Article{Gre90,
   author  = {L. Greengard},
   journal = {SIAM J. Sci. Stat. Comput.},
