### 2019-11-01 complexvariables/ComplexRegions.jl#4 @tobydriscoll
Tried to address these with 8553206.
