### 2019-11-01 complexvariables/ComplexRegions.jl#2 @tobydriscoll
I don't know what's causing the "connection refused" errors, though it's obviously in the Travis configuration that I set up using Documenter. It isn't happening with ComplexValues, so maybe comparing them will turn up something. Since the graphics were coming out anyway, I didn't make it a high priority.

If I leave the semicolons off of the plot commands, it adds a cryptic text message about the plot object to the output for the page. Do you know how to prevent that?
