### 2019-11-01 bae6a81
Trying to fix GKS issue with Travis docs construction.
