### 2019-11-01 66fdc75
Another example in the paper.
