### 2019-11-01 fc5a6f5
Removed semicolons from docs plot commands.
