+++ Project.toml
@@ -1,7 +1,7 @@
 name = "ComplexRegions"
 uuid = "c64915e2-6c82-11e9-38e9-1f159a780463"
 authors = ["Toby Driscoll <driscoll@udel.edu>"]
-version = "0.1.0"
+version = "0.1.1"
 
 [deps]
 Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
