+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"curves/#Curves-1","page":"Curves","title":"Curves","text":"","category":"section"},{"location":"curves/#","page":"Curves","title":"Curves","text":"using ComplexRegions,Plots\ndefault(linewidth=2,legend=:none)","category":"page"},{"location":"curves/#","page":"Curves","title":"Curves","text":"A curve is meant to be a smooth, non-self-intersecting curve in the extended complex plane. There is a generic Curve type that requires you to specify an explicit parameterization; it is not checked for smoothness or even continuity.","category":"page"},{"location":"curves/#interface_curves-1","page":"Curves","title":"Abstract interface","text":"","category":"section"},{"location":"curves/#","page":"Curves","title":"Curves","text":"Every AbstractCurve type is expected to implement the following methods. (Here C represents a value of type AbstractCurve and z is a number.)","category":"page"},{"location":"curves/#","page":"Curves","title":"Curves","text":"Method Description\npoint(C,t::Real) Complex point on C at parameter value t in [0,1].\ntangent(C,t::Real) Complex tangent to C at t.\nreverse(C) Reverse the direction of traversal.\nisfinite(C) True if the curve does not pass through infinity.\nconj(C) Complex conjugate of the curve.\nC+z Translate of the curve by z.\n-C Negate the curve.\nC*z Multiply the curve C by complex number z; i.e., scale and rotate it about the origin.\ninv(C) Invert the curve pointwise.","category":"page"},{"location":"curves/#","page":"Curves","title":"Curves","text":"There are also default implementations of the following methods:","category":"page"},{"location":"curves/#","page":"Curves","title":"Curves","text":"Method Description\npoint(C,t::AbstractArray{T<:Real}) Vectorization of the point method.\nz+C,C-z,z-C,z*C,C/z,z/C Translate/rotate/scale by a complex value.\nunittangent(C,t::Real) Normalized tangent to C at t.\nnormal(C,t::Real) Unit (leftward) normal to C at t.\narclength(C) Arc length of C.\nplotdata(C) Complex values that should be suitable for making a plot.","category":"page"},{"location":"curves/#","page":"Curves","title":"Curves","text":"There is also an AbstractClosedCurve subtype that is used to distinguish curves that close. It provides default implementations of the following methods.","category":"page"},{"location":"curves/#","page":"Curves","title":"Curves","text":"Method Description\nwinding(C,z::Number) Winding number of C about z.\nisinside(z::Number,C) Detect whether z lies inside the curve.\nisoutside(z::Number,C) Detect whether z lies outside the curve.","category":"page"},{"location":"curves/#generic_curves-1","page":"Curves","title":"Generic types","text":"","category":"section"},{"location":"curves/#Curve-1","page":"Curves","title":"Curve","text":"","category":"section"},{"location":"curves/#","page":"Curves","title":"Curves","text":"A Curve represents an implementation of AbstractCurve that requires only an explicit parameterization of the curve. Given the (bounded) complex-valued function f defined on 01, then C=Curve(f) represents the curve z=f(t). If f is defined on ab instead, then C=Curve(f,a,b) is appropriate, but all future work with C uses the standard interval 01 for the parameter. All Curve values are expected to be finite; i.e., isfinite(C) will always be true.","category":"page"},{"location":"curves/#","page":"Curves","title":"Curves","text":"By default a tangent to C is computed when needed using a simple finite difference, resulting in less precision than the representation of the points on C (particularly 
[…]
