### 2019-11-01 #1811 @dpsanders
There seems to be a problem with unicode characters in the Julia source code. @drvinceknight what can be done about that?
