### 2019-11-01 44198a4
Removed unicode and improved captions in the paper.
