### 2019-11-01 #1811 @tobydriscoll
I'm not married to using unicode for this submission. I've pushed a change that avoids it.
