### 2019-11-01 41298fd
Manual update of v0.1 docs
