### 2019-11-01 complexvariables/ComplexRegions.jl#2 @tobydriscoll
The GKS issue has been addressed, and I removed semicolons from the doc plotting commands.
