### 2019-11-04 #1811 @drvinceknight
@dpsanders and @dlfivefifty could you confirm you're happy with the requested modifications? (Thanks again for your time :+1:)
