### 2019-11-04 #1811 @dpsanders
The `References` heading in the PDF didn't come out right.
