### 2019-11-04 dd8b163
Blank line before References header
