### 2019-11-04 #1811 @tobydriscoll
@dpsanders I needed to put in a blank line. References header is ok now.
