### 2019-11-12 #1811 @drvinceknight
@dlfivefifty could you confirm you're now happy with the paper?

> Good for me, thanks.

@dpsanders I've taken the liberty to tick the two remaining boxes on your check list ("Community guidelines", "State of the field" and "Reference").
