### 2019-11-12 #1811 @dlfivefifty
Two comments:

1. The citation for IntervalSets.jl gives Huybrechs and myself (Olver) too much credit: we’ve made contributions but it was created by @mronian, with many contributions by @timholy
2. The use of “.jl” in package names should be consistent between “ComplexRegions” and “IntervalSets.jl and DomainSets.jl”.
