### 2019-11-14 d58a596
More paper tweaks
