### 2019-11-14 #1811 @tobydriscoll
Updates:

1. I have avoided trying to summarize credit for IntervalSets. The paper links to it, where the full history is easy to see.

2. I have decided to go without `.jl` on all the package names. It's an editorial choice, but seeing it over and over is really clunky.
