### 2019-11-14 #1811 @dlfivefifty
Ok one last comment: add references for DomainSets and IntervalSets as otherwise it’s not clear how to find them. This could just be links to the github pages
