### 2019-11-15 fde3a48
Moved refs in the paper
