+++ docs/paper.bib
@@ -216,3 +216,13 @@
 }
 
 @Comment{jabref-meta: databaseType:bibtex;}
+
+@misc{DomainSets,
+  title = {DomainSets package for Julia},
+  howpublished = {\url{https://github.com/JuliaApproximation/DomainSets.jl}}
+}
+
+@misc{IntervalSets,
+  title = {IntervalSets package for Julia},
+  howpublished = {\url{https://github.com/JuliaMath/IntervalSets.jl}}
+ }
