### 2019-11-15 #1811 @tobydriscoll
My mistake. I thought adding them as markdown links would get through. I put them in the Refs section instead.
