### 2019-11-15 aae3e29
Fixed the github refs in the paper
