+++ docs/paper.bib
@@ -218,11 +218,13 @@
 @Comment{jabref-meta: databaseType:bibtex;}
 
 @misc{DomainSets,
-  title = {DomainSets package for Julia},
-  howpublished = {\url{https://github.com/JuliaApproximation/DomainSets.jl}}
+  title = {DomainSets},
+  year = 2017,
+  howpublished = {{J}ulia package at \url{https://github.com/JuliaApproximation/DomainSets.jl}}
 }
 
 @misc{IntervalSets,
-  title = {IntervalSets package for Julia},
-  howpublished = {\url{https://github.com/JuliaMath/IntervalSets.jl}}
+  title = {IntervalSets},
+  year = 2016,
+  howpublished = {{J}ulia package at \url{https://github.com/JuliaMath/IntervalSets.jl}}
  }
