### 2019-11-16 #1811 @drvinceknight
@tobydriscoll it looks like a number of DOIs are missing, would you be able to add those please?

@dlfivefifty I believe the references you asked for are included now (but if you could confirm you're happy for this to be accepted that would be great :+1:)
