### 2019-11-16 #1811 @dlfivefifty
id Suggest adding “.jl” to the references, even if it’s not used in the paper itself: this seems more consistent with how things are cited, particularly with packages whose names otherwise might be confusing (DifferentialEquations.jl).

But I’ll leave it to the author/editor whether to incorporate this suggestion.
