### 2019-11-18 5c57637
Added DOIs to paper and renamed refs again
