+++ docs/paper.bib
@@ -6,6 +6,7 @@
   title   = {Conformal mapping of circular-arc polygons},
   year    = {1987},
   pages   = {19--32},
+  doi = {10/dnvcn7},
   volume  = {8},
   annote  = {First attempt at a full implementation of circular arcs. Partially successful.},
   groups  = {temp},
@@ -18,6 +19,7 @@
   year    = {1987},
   pages   = {1527--1530},
   volume  = {25},
+  doi = {10.2514/3.9819},
   annote  = {Computes lift on a plate in a tunnel and compares to standard approximations.},
   groups  = {temp},
 }
@@ -25,9 +27,10 @@
 @Article{DE93,
   author  = {T. K. De{L}illo and A. R. Elcrat},
   journal = {J. Comput. Phys.},
-  title   = {Numerical Conformal Mapping Methods for Exterior Regions with Corners},
+  title   = {Numerical conformal mapping methods for exterior regions with corners},
   year    = {1993},
   pages   = {199--208},
+  doi = {10.1006/jcph.1993.1175},
   volume  = {108},
   groups  = {temp},
 }
@@ -69,6 +72,7 @@
   year    = {1979},
   pages   = {236--239},
   volume  = {70},
+  doi = {10.1016/0022-247X(79)90086-6},
   annote  = {Capacitance is $-1/(4\pi)$ times integral of $\partial u/\partial n$ over one side, where $u$ is harmonic measure of opposite side. $C=(2/pi^2) \sum_{n=1,3,5,...} [n \sinh(n\pi\mu)]^{-1}$, where $\mu$ is the conformal modulus.},
   groups  = {temp},
 }
@@ -96,6 +100,7 @@
   year    = {1990},
   pages   = {603--620},
   volume  = {11},
+  doi = {10.1137/0911035},
   annote  = {Multipole solution of potential flow in a channel.},
   groups  = {temp},
 }
@@ -107,6 +112,7 @@
   year    = {1983},
   pages   = {287-307},
   volume  = {41},
+  doi = {10.1007/bf01418327},
   groups  = {temp},
 }
 
@@ -128,6 +134,7 @@
   year    = {1986},
   pages   = {125--142},
   volume  = {14},
+  doi = {10.1016/0377-0427(86)90133-0},
   groups  = {temp},
 }
 
@@ -138,6 +145,7 @@
   year    = {1992},
   pages   = {1836--1840},
   volume  = {40},
+  doi = {10.1109/22.156612},
   groups  = {temp},
 }
 
@@ -148,6 +156,7 @@
   year    = {1989},
   pages   = {475--487},
   volume  = {10},
+  doi = {10.1137/0910031},
   annote  = {Multipole for an integral equation.},
   groups  = {temp},
 }
@@ -170,6 +179,7 @@
   year    = {1986},
   pages   = {227--249},
   volume  = {14},
+  doi = {10.1016/0377-0427(86)90141-x},
   annote  = {Generalizes piecewise constant BCs to oblique derivatives. E.g., Hall effect, queueing.},
   groups  = {temp},
 }
@@ -181,6 +191,7 @@
   year    = {1982},
   pages   = {4980--4986},
   volume  = {53},
+  doi = {10.1063/1.331373},
   annote  = {See [Van83]. Here the shape is rectangular. The "Hall angle" makes this into an oblique-derivative problem.},
   groups  = {temp},
 }
@@ -218,13 +229,13 @@
 @Comment{jabref-meta: databaseType:bibtex;}
 
 @misc{DomainSets,
-  title = {DomainSets},
+  title = {DomainSets.jl},
   year = 2017,
[…]
