### 2019-11-18 #1811 @tobydriscoll
DOIs fixed and package names updated in the references.
