### 2019-11-20 #1811 @drvinceknight
Great, thanks.

It's all looking good to me now.

Could you make a tagged release and archive, and report the version number and archive DOI here (note that the archive title should match the paper title: `ComplexRegions.jl: A Julia package for regions in the complex plane`).
