+++ Project.toml
@@ -14,3 +14,8 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [compat]
 julia = "1"
+Colors = "^0.9"
+ComplexValues = "^0.2"
+RecipesBase = "^0.7"
+Reexport = "^0.2" 
+StaticArrays = "^0.12"
