### 2019-11-20 2dcfb37
Added DOI to readme
