### 2019-11-20 #1811 @tobydriscoll
It's v0.1.1, DOI is 10.5281/zenodo.3548866.

I hope it's all in order. Thanks!
