### 2019-11-20 #1811 @drvinceknight
@openjournals/joss-eics this paper is ready to be accepted :)

Thanks @dpsanders and @dlfivefifty again for your work reviewing the paper.
