+++ Project.toml
@@ -17,5 +17,6 @@ julia = "1"
 Colors = "^0.9"
 ComplexValues = "^0.2"
 RecipesBase = "^0.7"
-Reexport = "^0.2" 
+Reexport = "^0.2"
 StaticArrays = "^0.12"
+Distances = "^0.8"
