### 2019-11-21 c12ac7d
Revert "Bounded Distances dependency"
This reverts commit 5e6e0ce85a1efc5a3cad5ffc61a3fe99699c3b8e.
