### 2019-11-21 #1811 @kthyng
@tobydriscoll Hi! I am taking over from here. I made a few changes to your paper so there are not parentheses around references if they are meant to be inline instead. The PR is [here](https://github.com/complexvariables/ComplexRegions.jl/pull/5) and if you agree with it please merge.
