+++ docs/paper.md
@@ -17,9 +17,9 @@ bibliography: paper.bib
 
 # Summary
 
-Complex variables offer a unique and powerful way to represent planar regions and maps between them. Since the work of Riemann and others in the 19th century, complex variables have been used to derive quantities and properties of importance to a wide range of applications in science and engineering, in addition to their inherent mathematical importance. The advent of computing accelerated and extended this trend to applications in fluids, electromagnetics, queueing theory, computer graphics, and many other fields; e.g., @DE92;@Dri94;@Gai79;@Gre90;@MBH92;@TW86;@Ver82;@Gu2004.
+Complex variables offer a unique and powerful way to represent planar regions and maps between them. Since the work of Riemann and others in the 19th century, complex variables have been used to derive quantities and properties of importance to a wide range of applications in science and engineering, in addition to their inherent mathematical importance. The advent of computing accelerated and extended this trend to applications in fluids, electromagnetics, queueing theory, computer graphics, and many other fields; e.g., @DE92; @Dri94; @Gai79; @Gre90; @MBH92; @TW86; @Ver82; @Gu2004.
 
-Software packages for working with complex variables, particularly conformal maps, have been in the public domain for decades [@BG87;@Crowdy2016;@Dri96;@HP83;@Hu95;@Marshall2007;@Tre80]. There are additional papers based on research codes that apparently have not made it into the public domain, e.g. @DE93;@KT86;@OR89. These packages tend to take the traditional form of library routines that require some degree of facility with custom data structures used to represent geometry and functions. Moreover, computational methods are mostly specialized to particular types or classes of regions, leading to parallel and duplicated functionality within a package, and incompatible representations between them.
+Software packages for working with complex variables, particularly conformal maps, have been in the public domain for decades [@BG87; @Crowdy2016; @Dri96; @HP83; @Hu95; @Marshall2007; @Tre80]. There are additional papers based on research codes that apparently have not made it into the public domain, e.g. @DE93; @KT86; @OR89. These packages tend to take the traditional form of library routines that require some degree of facility with custom data structures used to represent geometry and functions. Moreover, computational methods are mostly specialized to particular types or classes of regions, leading to parallel and duplicated functionality within a package, and incompatible representations between them.
 
 The `ComplexRegions` package for Julia provides a software framework for the front end of computations over regions in the extended complex plane. It defines abstract types for curves, paths (i.e., piecewise smooth contours), and regions, defining minimal interfaces for them and providing default behavior based only on the interfaces. These are then implemented as concrete subtypes; e.g., a `Circle` is a subtype of `AbstractClosedCurve`, and a `Polygon` is a subtype of `AbstractClosedPath`. These types define data structures and methods to operate on them.
 
