### 2019-11-21 #1811 @kthyng
Also a [PR](https://github.com/complexvariables/ComplexRegions.jl/pull/6) for spacing I missed before.
