### 2019-11-21 #1811 @arfon
@kthyng - there are a couple of DOIs that are non-standard in format. There look to be alternative (standard-form) DOIs for the same records and I've updated them in https://github.com/complexvariables/ComplexRegions.jl/pull/7
