### 2019-11-22 #1811 @kthyng
@tobydriscoll Can you merge the PR from @arfon?

@arfon why are those errors only picked up when trying to accept instead of when generating the pdf?
