### 2019-11-23 #1811 @arfon
> @arfon why are those errors only picked up when trying to accept instead of when generating the pdf?

They're been generated when we create the Crossref metadata (which we don't do until towards the end of the process).
