### 2019-12-01 #1811 @Kevin-Mattheus-Moerman
I checked the archive on Zenodo and the paper and all looks good.
