### 2019-12-01 #1811 @Kevin-Mattheus-Moerman
@tobydriscoll it looks like you still need to merge this PR https://github.com/complexvariables/ComplexRegions.jl/pull/7. Let me know when you are ready to proceed.
