### 2019-12-02 #1811 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
